pyproject.toml
[tool.poetry]
name = "azure-openai-embedding-demo"
version = "0.1.0"
description = "A demo project for vector search using Qdrant with Azure OpenAI embeddings"
authors = ["Polin"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.12"  # 明確指定 Python 3.11.x
openai = "^1.12.0"
python-dotenv = "^1.0.0"
qdrant-client = "^1.12.0,<1.13"
numpy = "^1.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
